+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## TBA
 
+- Fixed: Document scan flow no longer proceeds to processing when additional documents are still selected locally.
 - `SDKInitRequestDataAttributes` now contains `platform` and `origin` properties (mainly to support BlinkID SDK).
 - Refactored document upload to use the new v2 API.
 - `ConfigurationService` allows to fetch Wultra Digital Onboarding configuration from the server.